重命名MarshoAI编写的bangumi代码
